bpf: use fake uname for bpfloader and netd Idea from here: CoreELEC/linux-amlogic@e761da6
This mechanism was used in the past to disable bpf on the Android userspace
faking an old kernel.

Starting with Android 14 QPR1 however Google will enforce kernel 4.19 to
just be able to boot Android.

In our case, with this 4.14 kernel, we already supports all the features
needed to boot Android 14 QPR1 (backported from 4.19): moreover, the userspace
checks again for the features available so it will not boot anyway in case
of issues.

For these reasons, we can provide to bpfloader, netbpfloader and netd a
fake kernel version to pass the kernel version checks.

Note that as now Android thinks we are 4.19, it also expects some features
from 4.19 that, as mentioned above, we already ported to 4.14.

Change-Id: I56936e128089b71bec197d7ba53a9de5071499ef
Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>
Signed-off-by: Shoiya Akari <barryofc11@gmail.com>